refactor : folder 구조 변경
