--- src/features/open-detail-post/model/useOpenDetailPost.tsx
@@ -1,14 +1,0 @@
-import { useAtom } from "jotai"
-import { openPostDetailAtom } from "./PostDetailContext"
-
-export const useOpenDetailPost = (fetchComments: (postId: number) => void) => {
-  const [, openPostDetail] = useAtom(openPostDetailAtom)
-
-  const handleOpenPostDetail = (post: any) => {
-    openPostDetail({ post, fetchComments })
-  }
-
-  return {
-    openPostDetail: handleOpenPostDetail,
-  }
-}
