--- src/features/open-edit-post/index.ts
@@ -1,2 +1,0 @@
-export { OpenEditPost } from "./ui/OpenEditPost"
-export { useOpenEditPost } from "./model/useOpenEditPost"
