+++ src/features/post/ui/OpenDetailPost.tsx
