+++ src/features/post/ui/RemovePostIcon.tsx
