--- src/features/post/ui/index.ts
@@ -1,3 +1,0 @@
-export { OpenDetailPost } from "../../open-detail-post/ui/OpenDetailPost"
-export { OpenEditPost } from "../../open-edit-post/ui/OpenEditPost"
-export { RemovePostIcon } from "../../remove-post-icon/ui/RemovePostIcon"
