--- src/features/remove-post-icon/index.ts
@@ -1,2 +1,0 @@
-export { RemovePostIcon } from "./ui/RemovePostIcon"
-export { useRemovePost, deletePostAtom } from "./model/useRemovePost"
