--- src/features/sort-by-select/data/index.ts
@@ -1,7 +1,0 @@
-// 정렬 기준 옵션 설정
-export const sortByOptions = [
-  { value: "none", label: "없음" },
-  { value: "id", label: "ID" },
-  { value: "title", label: "제목" },
-  { value: "reactions", label: "반응" },
-]
