--- src/features/sort-by-select/index.ts
@@ -1,2 +1,0 @@
-export { SortBySelect } from "./ui/SortBySelect"
-export { useSortBy } from "./model"
