--- src/features/sort-by-select/model/index.ts
@@ -1,1 +1,0 @@
-export { useSortBy } from "./useSortBy"
