--- src/features/sort-order-select/data/index.ts
@@ -1,5 +1,0 @@
-// 정렬 순서 옵션 설정
-export const sortOrderOptions = [
-  { value: "asc", label: "오름차순" },
-  { value: "desc", label: "내림차순" },
-]
