--- src/features/sort-order-select/index.ts
@@ -1,2 +1,0 @@
-export { SortOrderSelect } from "./ui/SortOrderSelect"
-export { useSortOrder, externalSortOrderAtom, setSortOrderAtom } from "./model/useSortOrder"
