+++ src/widgets/index.ts
@@ -1,3 +1,3 @@
-export { AddPostDialog } from "./post-dialog/AddPostDialog"
-export { EditPostDialog } from "./post-dialog/EditPostDialog"
-export { PostDetailDialog } from "./post-dialog/PostDetailDialog"
+export { AddPostDialog } from "./post-dialog/ui/AddPostDialog"
+export { EditPostDialog } from "./post-dialog/ui/EditPostDialog"
+export { PostDetailDialog } from "./post-dialog/ui/PostDetailDialog"
